requirements.txt
matplotlib==3.7.2
numpy==1.24.3
PyYAML==6.0.2
scipy==1.10.1
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1